pyproject.toml
# pyproject.toml

[build-system]

requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Locksmith"
version = "0.0.1"
description = "Locksmith – KERI identity vault application"
readme = "README.md"
requires-python = ">=3.12.6"
license = { text = "MIT" }

authors = [
    { name = "Phillip Feairheller", email = "pfeairheller@gmail.com" },
    { name = "Ari Argoud", email = "ariargoud@gmail.com" },
    { name = "Evan Asakawa", email = "evanja57@gmail.com" }
]

keywords = ["KERI Foundation", "cryptography", "KERI", "ACDC"]

classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "PySide6==6.9.2",
    "qasync>=0.28.0",
    "keri==1.3.4",
    "qtawesome>=1.4.0",
    "mnemonic==0.21",
    "multicommand==1.0.0",
    "pillow==11.0.0",
    "qrcode[pil]==7.4.2",
    "pyotp==2.9.0",
    "prometheus_client>=0.21.1",
    "schedule==1.2.2",
    "kerkle>=0.0.3",
    "cbor>=1.0.0",
    "falcon==4.0.2",
    "requests==2.32.3",
    'cryptography>=44.0.0',
    'pymongo>=4.11.2',
]

[project.entry-points."locksmith.plugins"]
# Example: to register a new plugin, add an entry like:
# myplugin = "locksmith.plugins.myplugin:MyPlugin"
# Then run: pip install -e .

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
pythonpath = ["src", "."]
